Update to VRCameraComponent
New movement handling which improves collision alot (still not working with slopes). Player can no longer stick their head through walls. Their is 2 options to handle when camera location is in an unsutable spot. Choice of the recenter the camera to the collision(snap camera back to the collision) or to just not let the camera move anymore.